+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 # Release History
 
+## 5.0.0-beta.3 [2019-10-09]
+
+- Updated: non live operation shouldn't stream large data, instead collect them all and zip as "large commands" instead.
+
 ## 5.0.0-beta.2 [2019-07-16]
 
-- Added eventLockId parameter in event.raise command.
-- Updated lodash dependency to 4.17.14.
+- Added: eventLockId parameter in event.raise command.
+- Updated: lodash dependency to 4.17.14.
 
 ## 5.0.0-beta.1 [2019-07-10]